requirements.txt
opencv_contrib_python==4.7.0.72
Pillow
PyYAML>=6.0.2
torch== 2.7.0+cu118
torchvision==0.22.0+cu118